add atomic_uint support for vc++

+++ include/Lock/SpinLock.h
@@ -159,18 +159,18 @@ namespace util
           std::atomic_uint m_enStatus;
 
         public:
-          SpinLock() : m_enStatus(Unlocked) {}
+          SpinLock() : m_enStatus(static_cast<unsigned int>(Unlocked)) {}
 
           void Lock()
           {
               unsigned char try_times = 0;
-              while (m_enStatus.exchange(Locked, std::memory_order_acq_rel) == Locked)
+              while (m_enStatus.exchange(static_cast<unsigned int>(Locked), std::memory_order_acq_rel) == Locked)
                   __UTIL_LOCK_SPIN_LOCK_WAIT(try_times ++); /* busy-wait */
           }
 
           void Unlock()
           {
-              m_enStatus.store(Unlocked, std::memory_order_release);
+              m_enStatus.store(static_cast<unsigned int>(Unlocked), std::memory_order_release);
           }
 
           bool IsLocked()
@@ -180,12 +180,12 @@ namespace util
 
           bool TryLock()
           {
-              return m_enStatus.exchange(Locked, std::memory_order_acq_rel) == Unlocked;
+              return m_enStatus.exchange(static_cast<unsigned int>(Locked), std::memory_order_acq_rel) == Unlocked;
           }
           
           bool TryUnlock()
           {
-              return m_enStatus.exchange(Unlocked, std::memory_order_acq_rel) == Locked;
+              return m_enStatus.exchange(static_cast<unsigned int>(Unlocked), std::memory_order_acq_rel) == Locked;
           }
 
         };
